POC: Dynamically add, edit or delete functions
The app is functional and can already plot functions.
It's possible to add a new function to the plot as well as edit or
delete an existing function.

The app is not robust, however. Exceptions of any kind will crash the
app, and the function uses `eval` to generate data, allowing for any
Python command to be passed and executed.