Improve page layout
Header and footer will be fixed to the top and bottom of the page.

The app will try to optimize the size of the graph.
It will try to occupy as much space as possible, changing the aspect
ratio as needed.
Nonetheless, the width and height of the graph are bounded to the
following limits:

- width: ≥350px and ≤1600px
- height:≥350px and  ≤800px