Implement session separation
Depending on the browser, different instances of the app can be served
in parallel, allowing different graphs to be opened.

The actual behavior is browser-dependent.
E.g., in Safari, each "private window/tab" will be identified as a
independent session, while in Opera, all "private windows/tabs" are part
of the same session.